Se añade el script para borrar la papelera.

+++ Scripts/clean_trash.sh
@@ -1,9 +1,14 @@
 #!/bin/bash
 
-echo "Accediendo a la papelera..."
-cd $HOME/.local/share/Trash/files/
-echo "$(pwd)"
-echo "Presiona una tecla para continuar, CTRL+C para salir..."
-read
-echo "Eliminando archivos..."
-rm -rf *
+echo "Moving to trash directory..."
+RUTA_PAPELERA="$HOME/.local/share/Trash/files"
+
+if [ -d $RUTA_PAPELERA ]
+then
+	cd $RUTA_PAPELERA
+	echo "Moviendo al directorio: $(pwd)"
+	rm -rf *
+	echo "Se han eliminado todos los archivos."
+else
+	echo "No se encuentra el directorio de la papelera..."
+fi
